+++ packages/admin/src/mui/menu/Context.tsx
@@ -3,6 +3,7 @@ import * as React from "react";
 export interface IMenuContext {
     open: boolean;
     toggleOpen: () => void;
+    headerHeight: number;
 }
 
 export interface IWithMenu {
@@ -14,6 +15,7 @@ export const MenuContext = React.createContext<IMenuContext>({
     toggleOpen: () => {
         // nothing
     },
+    headerHeight: 0,
 });
 
 export const withMenu = <P extends object>(WrappedComponent: React.ComponentType<P & IWithMenu>): React.FunctionComponent<P> => {