+++ packages/admin/src/themeAugmentation.ts
@@ -1,9 +1,11 @@
 import { ClearInputButtonThemeProps, CometAdminClearInputButtonClassKeys } from "./common";
 import { CometAdminFormFieldContainerClassKeys, CometAdminInputBaseClassKeys, FieldContainerThemeProps } from "./form";
 import {
+    CometAdminMasterLayoutClassKeys,
     CometAdminMenuClassKeys,
     CometAdminMenuCollapsibleItemClassKeys,
     CometAdminMenuItemClassKeys,
+    MasterLayoutThemeProps,
     MenuCollapsibleItemThemeProps,
     MenuThemeProps,
 } from "./mui";
@@ -15,6 +17,7 @@ declare module "@material-ui/core/styles/overrides" {
         CometAdminMenu: CometAdminMenuClassKeys;
         CometAdminMenuItem: CometAdminMenuItemClassKeys;
         CometAdminMenuCollapsibleItem: CometAdminMenuCollapsibleItemClassKeys;
+        CometAdminMasterLayout: CometAdminMasterLayoutClassKeys;
         CometAdminInputBase: CometAdminInputBaseClassKeys;
     }
 }
@@ -25,5 +28,6 @@ declare module "@material-ui/core/styles/props" {
         CometAdminFormFieldContainer: FieldContainerThemeProps;
         CometAdminMenu: MenuThemeProps;
         CometAdminMenuCollapsibleItem: MenuCollapsibleItemThemeProps;
+        CometAdminMasterLayout: MasterLayoutThemeProps;
     }
 }